refactor(fbuild-cli): split main.rs into submodules to satisfy LOC gate
