+++ crates/fbuild-cli/src/README.md
@@ -2,6 +2,8 @@
 
 ## Modules
 
-- **`main.rs`** -- CLI entry point; Clap parser with subcommands (build, deploy, monitor, reset, purge, daemon, device, show, mcp, clang-tidy, iwyu, clang-query), dispatches to daemon client
+- **`main.rs`** -- CLI entry point; spawns the larger-stack `fbuild-main` thread that hands control to `cli::async_main`
+- **`cli/`** -- Clap parser, subcommand argument types, and per-subcommand handlers (see `cli/README.md`); split out of `main.rs` to keep each `.rs` under the 900 LOC gate
+- **`mcp/`** -- MCP (Model Context Protocol) stdio JSON-RPC server (see `mcp/README.md`); split out of a flat `mcp.rs` to keep each `.rs` under the 900 LOC gate
 - **`daemon_client.rs`** -- `DaemonClient` HTTP client, request/response types, daemon spawn with stale binary detection, NDJSON streaming, compact status display
-- **`mcp/`** -- MCP (Model Context Protocol) stdio JSON-RPC server; translates AI assistant tool/resource calls into daemon HTTP requests. Exposes `mcp::run_mcp_server` (see `mcp/README.md` for the per-file breakdown).
+- **`lib_select.rs`** -- diagnostic LDF-style library selection resolver (`fbuild lib-select`)
